move: note details page and button to common dir
