+++ src/components/NotesMindMap/NotesMindMapComponent.js
@@ -6,7 +6,7 @@ import { VisNetworkHelper } from 'helpers/visNetworkHelper';
 import noteStorage from 'storage/noteStorage';
 import { NoteNameEditorComponent } from 'components/NoteNameEditor/NoteNameEditorComponent';
 import { StyledNotesMindMap } from 'components/NotesMindMap/NotesMindMapStyles';
-import { NoteDetailsButtonComponent } from 'components/NoteDetailsButton/NoteDetailsButtonComponent';
+import { NoteDetailsButtonComponent } from 'components/NoteDetails/NoteDetailsButtonComponent';
 
 export class NotesMindMapComponent extends Component {
   render() {
@@ -81,7 +81,6 @@ export class NotesMindMapComponent extends Component {
 
     return (
       <StyledNotesMindMap>
-        {/* {selectedNote && <NoteDetailsContainer />} */}
         {selectedNote && <NoteDetailsButtonComponent />}
 
         <VisGraph graph={visGraph} events={visEvents} options={visOptions} />
